Make ExternalCommandHelper::getNonce() reentrant.
Store previously generated values of nonce, and remove them from
the container when they are used.

BUG: 396312

+++ src/util/externalcommand.h
@@ -130,13 +130,13 @@ private:
 
     void onFinished(int exitCode, QProcess::ExitStatus exitStatus);
     void onReadOutput();
-    static unsigned long long getNonce(QDBusInterface& iface);
+    static quint64 getNonce(QDBusInterface& iface);
 
 private:
     std::unique_ptr<ExternalCommandPrivate> d;
 
-    // KAuth stuff
-    static unsigned long long m_Nonce;
+    // KAuth
+    static quint64 m_Nonce;
     static KAuth::ExecuteJob *m_job;
     static QCA::Initializer *init;
     static QCA::PrivateKey *privateKey;
